requirements.txt
bilibili-api-python>=16.0.0
httpx>=0.24.0
